+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PixDispute resource
 - PixDispute.Transaction sub-resource
 - PixDispute.Log sub-resource
+- priority and reason attributes and response method to PixRequest resource
 ### Changed
 - IssuingBin resource to IssuingProduct
 - settlement parameter to funding_type and client parameter to holder_type of IssuingProduct